feat: implementando a remoção de um usuario

+++ app/src/main/java/devandroid/micaela/tcc_agendamed/controller/UsuarioController.java
@@ -33,7 +33,7 @@ public class UsuarioController{
         return this.colecaoUsuarios.editar(usuario);
     }
 
-    public int remover(int id) {
+    public int remover(long id) {
         return this.colecaoUsuarios.remover(id);
     }
