+++ app/src/main/java/devandroid/micaela/tcc_agendamed/model/ColecaoUsuarios.java
@@ -6,6 +6,6 @@ public interface ColecaoUsuarios {
     long inserir(String nome);
     public List<Usuario> obterTodos();
     public int editar(Usuario usuario);
-    public int remover(int id);
+    public int remover(long id);
 
 }
